worker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
celery[redis]==5.4.0
pydantic-settings==2.7.1
httpx==0.28.1
google-genai==1.7.0
weaviate-client==4.10.4
